Move isEarlierSemester and getStartingPlanSemester from plannerUtils.ts => utilFunctions.ts
