+++ src/utils/utilFunctions.ts
@@ -3,7 +3,6 @@ import { SemesterCode } from 'prisma/utils';
 import { UUID } from 'bson';
 
 import { Semester } from '@/components/planner/types';
-import { isEarlierSemester } from './plannerUtils';
 import { tagColors } from '@/components/planner/utils';
 
 /**
@@ -174,4 +173,32 @@ const regex =
 
 export const isValidEmail = (email: string) => {
   return regex.test(email);
+};
+
+/**
+ * Is semesterOne earlier than semesterTwo
+ */
+export const isEarlierSemester = (semesterOne: SemesterCode, semesterTwo: SemesterCode) => {
+  if (isSemCodeEqual(semesterOne, semesterTwo)) {
+    return false;
+  } else if (semesterOne.year > semesterTwo.year) {
+    return false;
+  } else if (
+    semesterOne.year === semesterTwo.year &&
+    (semesterOne.semester === 'f' || (semesterOne.semester === 'u' && semesterTwo.semester === 's'))
+  ) {
+    return false;
+  }
+  return true;
+};
+
+export const getStartingPlanSemester = (): SemesterCode => {
+  const d = new Date();
+  if (d.getMonth() < 5) {
+    return { year: d.getFullYear(), semester: 's' };
+  } else if (d.getMonth() > 7) {
+    return { year: d.getFullYear(), semester: 'f' };
+  } else {
+    return { year: d.getFullYear(), semester: 'u' };
+  }
 };
